Validation changes and phone_number add
